# Conflicts:
#	ClickyDicky/Assets/Scenes/GameScene.unity
#	ClickyDicky/Assets/Scripts/Clicky/MoneyClicker.cs
#	ClickyDicky/Assets/Scripts/NyarLog.meta
#	ClickyDicky/Library/CurrentLayout.dwlt
#	ClickyDicky/Library/ScriptAssemblies/Assembly-CSharp.dll
#	ClickyDicky/Library/ScriptAssemblies/Assembly-CSharp.dll.mdb
#	ClickyDicky/Library/assetDatabase3
#	ClickyDicky/Library/expandedItems
#	ClickyDicky/Library/metadata/00/00000000000000002000000000000000
#	ClickyDicky/Library/metadata/00/00000000000000003000000000000000
#	ClickyDicky/Library/metadata/00/00000000000000004000000000000000
#	ClickyDicky/Library/metadata/00/00000000000000004100000000000000
#	ClickyDicky/Library/metadata/00/00000000000000005000000000000000
#	ClickyDicky/Library/metadata/00/00000000000000005100000000000000
#	ClickyDicky/Library/metadata/00/00000000000000006000000000000000
#	ClickyDicky/Library/metadata/00/00000000000000006100000000000000
#	ClickyDicky/Library/metadata/00/00000000000000007000000000000000
#	ClickyDicky/Library/metadata/00/00000000000000007100000000000000
#	ClickyDicky/Library/metadata/00/00000000000000008000000000000000
#	ClickyDicky/Library/metadata/00/00000000000000008100000000000000
#	ClickyDicky/Library/metadata/00/00000000000000009000000000000000
#	ClickyDicky/Library/metadata/00/0000000000000000a000000000000000
#	ClickyDicky/Library/metadata/00/0000000000000000a100000000000000
#	ClickyDicky/Library/metadata/00/0000000000000000b000000000000000
#	ClickyDicky/Library/metadata/00/0000000000000000c000000000000000
#	ClickyDicky/Library/metadata/02/02e08a006ecfb364684a36d657e4c347
#	ClickyDicky/Library/metadata/02/02e08a006ecfb364684a36d657e4c347.info
#	ClickyDicky/Library/metadata/3c/3c0ad459c1534645b5d603b7cc258f97
#	ClickyDicky/Library/metadata/3c/3ca1789553e464b4a85406a78e23101b
#	ClickyDicky/Library/metadata/45/4501a059a7b177744bab86a31ef1ffc4
#	ClickyDicky/Library/metadata/5f/5f32cd94baa94578a686d4b9d6b660f7
#	ClickyDicky/Library/metadata/73/739bbd9f364b4268874f9fd86ab3beef
#	ClickyDicky/Library/metadata/7c/7cbab2be89b54486bbd23a6fe637d30e
#	ClickyDicky/Library/metadata/80/80a3616ca19596e4da0f10f14d241e9f
#	ClickyDicky/Library/metadata/85/852e56802eb941638acbb491814497b0
#	ClickyDicky/Library/metadata/87/870353891bb340e2b2a9c8707e7419ba
#	ClickyDicky/Library/metadata/8e/8e0cd8ed44d4412cbe0642067abc9e44
#	ClickyDicky/Library/metadata/8f/8f136a6e7fb8b4d47a2b86e5fe2ae26d
#	ClickyDicky/Library/metadata/97/97decbdab0634cdd991f8d23ddf0dead
#	ClickyDicky/Library/metadata/a5/a543bd2f47ec0f94f9806b477107e4e1
#	ClickyDicky/Library/metadata/ad/adebbd281f1a4ef3a30be7f21937e02f
#	ClickyDicky/Library/metadata/f2/f224dcc0890c639409cf1e3fef8eacfd
#	ClickyDicky/Library/metadata/f4/f47b476f278f0cd4ea57714a02dc71ff
#	ClickyDicky/Library/metadata/f5/f5f67c52d1564df4a8936ccd202a3bd8

+++ ClickyDicky/Assets/Scripts/Game Manager/GameManager.cs
@@ -8,7 +8,7 @@ public class GameManager : MonoBehaviour
     public static GameManager manager;
 
     [Header("Money Info")]
-    public int moneyAvaliable;
+    public float moneyAvaliable;
 
     #region Initialisation
     void Awake()
